+++ src/db_sub_source.hh
@@ -117,6 +117,7 @@ public:
     nonstd::optional<size_t> dls_time_column_invalidated_at;
     std::unique_ptr<ArenaAlloc::Alloc<char>> dls_allocator{
         std::make_unique<ArenaAlloc::Alloc<char>>(64 * 1024)};
+    string_attrs_t dls_ansi_attrs;
 
     static const char NULL_STR[];
 };